TileList extends ArrayList<Tile>
TileList now extends ArrayList<Tile> instead of being retarded.
Implemented. Fixed some ambiguities in HandChecker.

+++ src/majava/HandChecker.java
@@ -1114,7 +1114,7 @@ public class HandChecker {
 				checkTilesMinusThisMeld = checkTiles.makeCopyWithCheckers();
 				toMeldTiles = new TileList();
 				
-				while (!partnerIndices.isEmpty()) toMeldTiles.add(checkTilesMinusThisMeld.remove(partnerIndices.remove(partnerIndices.size() - 1)));
+				while (!partnerIndices.isEmpty()) toMeldTiles.add(checkTilesMinusThisMeld.remove( partnerIndices.remove(partnerIndices.size() - 1).intValue()) );
 				toMeldTiles.add(checkTilesMinusThisMeld.removeFirst());
